+++ football_love/src/main/java/com/deu/football_love/repository/CompanyRepository.java
@@ -1,29 +1,12 @@
 package com.deu.football_love.repository;
 
 import com.deu.football_love.domain.Company;
-import com.deu.football_love.domain.Team;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Repository;
 
-import javax.persistence.EntityManager;
+import java.util.List;
 
-@Repository
-@RequiredArgsConstructor
-public class CompanyRepository {
-    private final EntityManager em;
-
-    public Company selectCompany(Long companyId)
-    {
-        return em.find(Company.class, companyId);
-    }
-
-    public void insertCompany(Company company)
-    {
-        em.persist(company);
-    }
-
-    public void deleteCompany(Company company)
-    {
-        em.remove(company);
-    }
+public interface CompanyRepository {
+    Company selectCompany(Long companyId);
+    List<Company> selectCompaniesById(String companyName);
+    Company insertCompany(Company company);
+    void deleteCompany(Company company);
 }
